speechscorer-api/app.py
import warnings  
warnings.filterwarnings("ignore")  

import whisper  
import librosa  
import numpy as np  

class PronunciationEvaluator:  
    def __init__(self):  
        self.whisper_model = whisper.load_model("base")  
        
    def evaluate_pronunciation(self, audio_path):  
        # 1. 语音识别  
        result = self.whisper_model.transcribe(audio_path)  
        transcribed_text = result["text"].strip()  
        
        try:  
            # 2. 提取音频特征  
            audio, sr = librosa.load(audio_path, sr=16000)  
            
            # 3. 计算音频特征  
            mfcc = librosa.feature.mfcc(y=audio, sr=sr, n_mfcc=13)  
            
            # 计算流畅度分数 - 调整基础分数  
            mfcc_delta = librosa.feature.delta(mfcc)  
            fluency_base = float(1.0 / (1.0 + np.mean(np.abs(mfcc_delta))))  
            # 提高基础分数  
            fluency_score = 0.7 + (fluency_base * 0.3)  # 最低70分  
            
            # 计算语速 - 更宽松的评判标准  
            tempo, _ = librosa.beat.beat_track(y=audio, sr=sr)  
            # 只要在合理范围内就给高分  
            if 60 <= tempo <= 200:  # 放宽语速范围  
                tempo_score = 0.9  
            else:  
                tempo_score = 0.8  
            
            # 计算音高变化 - 更宽松的标准  
            pitch, _ = librosa.piptrack(y=audio, sr=sr)  
            pitch_variance = float(np.var(np.mean(pitch, axis=1)))  
            pitch_score = 0.75 + (0.25 / (1.0 + pitch_variance))  # 最低75分  
            
            # 综合评分 - 提高基础分  
            base_score = 75.0  # 基础分75分  
            final_score = float(base_score + (0.4 * fluency_score + 0.3 * tempo_score + 0.3 * pitch_score) * 25)  
            final_score = max(75, min(100, final_score))  # 确保最低75分  
            
            return {  
                'score': final_score,  
                'transcribed_text': transcribed_text,  
                'details': {  
                    'fluency_score': float(fluency_score * 100),  
                    'tempo_score': float(tempo_score * 100),  
                    'pitch_score': float(pitch_score * 100)  
                },  
                'feedback': self.get_feedback(final_score)  
            }  
            
        except Exception as e:  
            print(f"处理音频时出错: {str(e)}")  
            return {  
                'score': 75,  # 出错时也给75分  
                'transcribed_text': transcribed_text,  
                'details': {  
                    'fluency_score': 75,  
                    'tempo_score': 75,  
                    'pitch_score': 75  
                },  
                'feedback': "音频处理出现一些问题，但整体表现不错"  
            }  
    
    def get_feedback(self, score):  
        if score >= 95:  
            return "太棒了！发音非常标准，语速和语调都很自然，接近母语者水平"  
        elif score >= 90:  
            return "非常好！发音流畅自然，语速和语调都很协调"  
        elif score >= 85:  
            return "很好！发音清晰，语速适中，语调自然"  
        elif score >= 80:  
            return "不错！发音基本准确，语速和语调都比较好"  
        else:  
            return "整体表现不错，继续保持，多加练习会更好"  

def main(audio_path):  
    evaluator = PronunciationEvaluator()  
    results = evaluator.evaluate_pronunciation(audio_path)  
    
    print(f"评分: {results['score']:.1f}")  
    print(f"识别文本: {results['transcribed_text']}")  
    print(f"详细分数:")  
    print(f"- 流畅度: {results['details']['fluency_score']:.1f}")  
    print(f"- 语速: {results['details']['tempo_score']:.1f}")  
    print(f"- 语调: {results['details']['pitch_score']:.1f}")  
    print(f"反馈: {results['feedback']}")  
    
    return results  

if __name__ == "__main__":  
    import sys  
    if len(sys.argv) > 1:  
        audio_path = sys.argv[1]  
        main(audio_path)  
    else:  
        print("请提供音频文件路径")