speechscorer-api/requirements.txt
openai-whisper
torch
transformers
librosa
jiwer
dtaidistance
edge-tts